Update project to current development state

+++ src/Songbook.Web/Pages/Songs/Delete.cshtml.cs
@@ -1,11 +1,14 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
+using Songbook.Web.Auth;
 using Songbook.Web.Data;
 using Songbook.Web.Models;
 
 namespace Songbook.Web.Pages.Songs;
 
+[Authorize]
 public class DeleteModel : PageModel
 {
     private readonly SongbookDbContext _context;
@@ -26,6 +29,11 @@ public class DeleteModel : PageModel
         if (Song == null)
             return NotFound();
 
+        var isAdmin = User.IsInRole("Admin");
+        var myProfileId = User.GetProfileId();
+        if (!isAdmin && Song.CreatedByUserId != myProfileId)
+            return Forbid();
+
         return Page();
     }
 
@@ -34,6 +42,11 @@ public class DeleteModel : PageModel
         var song = await _context.Songs.FindAsync(id);
         if (song != null)
         {
+            var isAdmin = User.IsInRole("Admin");
+            var myProfileId = User.GetProfileId();
+            if (!isAdmin && song.CreatedByUserId != myProfileId)
+                return Forbid();
+
             _context.Songs.Remove(song);
             await _context.SaveChangesAsync();
         }
